Initial Commit
Added two files `anagram.py` and `funcs.py` to organise code efficiently.

The anagram file
This is the main program and from here, all functions can be accessed

The func file
This is a class file that contains all the game logic such as selecting
a word at random, shuffling the word, guessing the correct word etc.

The .gitignore file
Updated the content of this file for git to ignore unneeded files