Code Improvement
Improved function efficiency

The funcs file
In the shuffle function, I commentted the line that obtains a word by
calling the get_word() function. Then, I added a parameter to the
function's definition. This is made to ensure that, the get_word()
function will be called once in the main program and its value passed to a
variable.