Code Improvement
Added a new function, guess_word()

The guess_word function
This function invokes the get_word() and shuffle_word() functions. The
get_word function picks a word from the file and passes it to the
shuffle_word function which shuffles the word. The guess_word then
presents the shuffled word to the player to guess the correct word. An
input by the player is compared with the word and if it matches, the
function prints "congrats" and returns 1. It returns -1 otherwise.