Code Improvement
Did several changes to functions to improve application logic

The funcs File
Changed the guessword() function to accept only input from the user.
This function does not display the shuffled word.

Improved the getscore() function to invoke the getword(), shuffle_word()
and guessword() functions to compare the user's input to the actual
word. The user has 3 attempts to guess the correct word. If at any
moment the user guesses the correct answer, the program breaks out of
the loop and awards a score of 3. If after the 3rd attempt the user is
unable to supply the correct word, a score of 0 is awarded.