Code Improvement
Added several lines to improve the application's functionality as
described below

anagram.py file
In the main file, added four lines to display at the top the mode of the
application, commands to use to navigate the game such shuffle the word
(s), pass a guess (Enter/Return) or quit the game (q).

funcs.py file
In the funcs file, added two more elif block statements to execute the
commands to shuffle the word, pass a guess or quit the game mode.

Several test runs updated the contents of the database file