New Class File
Added a new class file, category.py. This file is however temporary; it is
still in the early stage of development. The file contains the usual
init function to initialise the class a with file path as a parameter.
The other methods, _getfilepath returns a path to the database file;
_savenewcategory saves a new category data to the database.