Added a New Class
dbclass.py is a database management module with methods for creating a
table in a specified database, to commit data to the database, select
records for display, delete records from the database among other
functionalities. This module was originally created for a previous
project I did and it will be modified to suit the current project as
development progresses.