Added Methods to Display Records
Added a new function to display records and updated other methods to
reflect the new changes

dataentry.py file
Added a new method, _getrecords() that executes in a loop until a
condition is met at which it terminates. This method invokes several
methods from the dbclass.py module to display all sorts of records.

dbclass.py file
Added a new method, display_list_records() that selects detail records
from a specified table. At the moment, the only record it displays is
the Categories information. A note in comments has been added to further
improve on this feature. Given the different records to display, this
method will invoke another module yet to be developed to create a
specific table (Texttable) for the display of a detail list of a
specified record.

mywallet.db file
The program was ran several times and debugged which updated the
contents of the database file

mywallet.py file
A new line was added to the switch block to invoke the new method
described in the dataentry.py file. This displays the records on the
screen.