Fixed Database Related Bug
Selecting records from the database to be displayed in a combobox widget
on a gui was problematic. However, I fixed it by updating two separate
files as described below:

app/dbclass.py
This file has a method that selects record names from the database.
Initially, it prints the records in the console with the help of a
Texttable module. I commented out the lines that does this just returned
the variable to which recordsi are to passed after fetching them from the
database.

newincome.py file
The main function of this module is a gui and has a combobox widget
in which lists of item names from the database will be displayed. To
achieve this, I created a function name getrecords() which creates a
list with the names recieved from the database. The initial problem was
I did not pass the path to the databse file to the class that manages
the database. Passing the database file path to this class solved the
problem. The record names from the database now display without any
error in the combobox widget on the gui.