File reorganisation
Reorganised files to make it easier to import them among other files. In
the course of development, I was faced with an import situation that was
near impossible to import files from parent directory. Although I could
implement other means to work around, I've chosen to rearrange the
files:

The main module with the main file is the parent directory. In this
directory is the gui directory in which are modules for modeling various
interfaces of the application. In the gui directory is the application
logic directory which contains the backend logic of the gui files. In
this directory is the appdb directory which contains the database file and the
dbclass module file. In this arrangement, importing is done from
directory to parent directory and not the other way round. This is
because modules in parent directories depend on the modules in the child
directories. In short, the file arrangement is as follows:

app parent directory->gui directory->applogic directory->appdb directory