Update in Newcategory Module
In the category module, I moved the function close() to the bottom of
the file. This is done to follow the principle that states, all
functions that do not call other functions should be at the bottom of
the file. On the other hand, functions that call other functions are
placed at the top region in a module. This arrangement presents a clean
layout for functions and methods and facilitates editing and debugging.