Added Two Functions
I worked on taking data from the gui. For some reason I am yet to know,
the variables assigned to the gui widgets cannot be identified and so
the program cannot fetch the data from the gui. Below are the additions
I made to the functions

newcatgui.py module
In this module, I added a new function saveData() that fetches data
entered by the user into the gui. From the tutorials I found around the
web, I can use the get() method from tkinter to get the data entered
into the gui. However, the compiler applies a yellow wiggly line under
each of the variables assigned to the gui widgets with the message that,
the variables are not defined when I applied the get() method to them in
the saveData() function. I will work around to solve this. It invokes
the saveCategoryData() from the applogic module and passes on the
collected data to be committed to the database.

applogic.py module
In this module, I updated the saveCategoryData() function to reflect the
changes made in the previous file which invokes this function. At the
moment, this function does not commit the data to the database; it
prints a confirmation message to signify that data transfer between the
two modules was a success.