Solved Variable Not Found Problem
In the previous commit, I explained that applying the get() method from
tkinter to variables assigned to the gui widgets was not working as the
compiler applied yellow wiggly lines under them with the message that,
they are not defined.

I tried multiple ways by placing the function at the bottom and the top
of the module at different times. At another attempt, I observed that
the definitions for the gui widgets were in a function named main(). So I
removed the definitions from the function and placed them as is in the
file and just then, the wiggly lines disappeared. I ran the program
several times after some adjustments and viola, it produces the messgae
I expected.

On this note, I am fully certain that data is successfully transferred
from one module to the other. I will now proceed to code for the
saveCategoryData() function in the applogic module that in turn invokes
_commitcategory() function from the dbclass module to actually commit
the data to the database.