Changed method name, added new mehtod and comments
new method
Added the prototype of a new function, messages() to be used to report
messages to the user. The message will appear on the gui in red text to
contrast the grey background and make it more visible to the user. Note
that the name of this method is subject to change.

method name change
Changed method's name from saveData() to aggregateData(). This was done
to reflect what this method actually does; collect data entered by the
user from the gui widgets

comments
Added comments to the following methods: messages() and aggregateData()