New features and improvement
This commit covers the following activities as explained in detail
later: Changed gui size, added a new method, updated another method and
improved comments

Changed gui size
In order to accommodate the new feature to be added which was a Label
widget for displaying messages at the bottom region of the gui frame, I
increased the height from 210 to 230.

Added a new method
As stated in the previous commit, I added the logic for the new method
and changed its name from messages() to displayMessages() to reflect
what it actually does. This method, when invoked receives an integer
value to determine the message to display and the color for the texts.
It then creates a new Label widget and adds it to the grid container of
the gui frame.

Updated another method
I updated the aggregateData() method by invoking the displayMessages()
method and passing an integer coded value to it based on the type of
message to be displayed. The types of messages it invokes are invalid
entries and the success or failure of saving a data to the database.

Improved comments
Improved the comments of all affected methods, and fixed a few typos as
well.