Error Message Improvement
After the success of displaying error messages, I added code to check
for invalid entries; so far the program is able to detect empty and
invalid entries for the entry and the combobox widgets, yet to discover
 how to do same for the text widget.

As I ran the program several times, I realised new messages do not
override old messages; the new mesage appears to sit ontop of the old message
 and if the new message is shorter, the old message can still be seen
underneath the new message. This is the current stage I am working to
fix.