Code Improvement
Finally, I have managed to fixed the problem associated with old
messages appearing under new messages. Initially, I set the sticky
property of the label which displays the messages to tk.W, thus sticking
to the left handside. I only changed the sticky property to
tk.N+tk.S+tk.W+tk.E which causes the message to appear centered in the
label because this property setting causes the label to stick to all
borders of the column it appears in the grid container.

I am yet to discover how to check for empty values in the Text widget.
I'm really enjoying this process; I'm learning more than I can imagine.