New Program Features
Worked on applogic.py and newincome.py modules as explained below:

newincome module
This module has simialr functionality with the new category module and
so I copied some methods from the new category module. The methods
copied were displayMessages() and the aggregateData().

For the aggregateData() method, I renamed it to aggregateIncomeData() to
be specific with this module and made some edits in the documentation
comments. Then, I updated the variables and function calls to suit the
activity of this module.

The only challenge in this module is checking to ensure the user enters
integer or float values in the amount entry widget. I can typecast the
assigned variable to float type at the point of taking the data from the
widget but this can crash the program should the user enter string type and
click on the save button. For this reason, a default value of 0.0 has
been set to vividly instruct the user to enter numeric values only. As
it can been in this method, the elif block that checks for the type of
the amount widget has been commented or else even numeric values entered
will be considered as string type as outputs from entry widgets are of
type string

For the displayMessages() method, the name was maintained as is but the
variables and associated messages were updated to suit the function of
the income module.

applogic module
For this module, I updated the saveIncomeData() method with parameters
to receive data from the new income module. Also, I updated the messages
it prints to reflect the data entered by the user in the income gui.

other modifications
In the new income module, the combo widget that displays a list of
category names from the database has been deactivated and supplied with a
dummy list of categories defined in the module. This is because the
database does not have any records at the moment. The function call to
this widget has been commented at the end of the line to be used later
when the database is populated with records.