Database Connection
Connected the backend to the database and several runs executes
successfully. Below is the update

newcatgui module
Initially, I planned to use a separate file to handle the backend logic
for the gui applications. In this file will be the method to connect to
the module that connects to the database. However, I skipped over that
backend logic file and invoked the method from the dbclass module
directly in the gui module to commit the data to the database and this
works well.

dbclass module
In this module, I updated the document comment, and the sql statement

mywallet.db
This is the database file. Successful program execution updated the
content of the categories table with data from the gui widgets