Updated Code in New Income GUI
Previously, I commented out the getCategoryRecords() method from the
applogic module which is assigned to the combobox widget on the new
income gui. After successfully connecting and saving category data to
the database, I reassigned the method to the combobox widget and when
program executes, the category types display in the combobox.

Although connection works as expected, it fetches the wrong data and
this is due the backend logic for the new category gui; it saves wrongly
arranged data which will be fixed in the next commit.