Fixed Wrong Data Selection
Quickly fixed the wrong data selection problem explained in the
immediate previous commit.

The problem was actually in the commit category method of the dbclass
where I mistakingly interchanged the positions of two parameters data
tuple set. After fixing this up, the program runs, successfully inserts
new data into the database and successfully displays the category types
in the combobox widget of the new income gui.