New Record
After creating a method to commit new income record to the database in
the dbclass module, I invoked this method in the new income gui module
and passed the data from the gui widgets to the method parameters. When
executed, the program runs without any error and saves the record to the
database.

Also, I commentted out the line that invokes methods from the not-in-use
applogic module.