Code Improvement
Initially, I had no use of the applogic file but finally found it
useful:

dbclass module
In this module, I was supposed to create a new method to fetch categories
of type expense from the categories table in the database. However, I
realised I could improve the display_list_records methods which displays
a list of categories from the categories table. The improvement is to
add a parameter which will determine the category type to fetch records;
this prevented the need to duplicate this method purposely to fetch
records of another type. I then added the parameter to the sql statement
which will fetch only records of a type passed to the parameter.

newincomegui module
The above update will require that an argument be passed to the
function call in the new income gui module. The method that invokes the
display_list_method is the getCategoryRecords() in this module. However,
I realised I would have to duplicate this method for the new expense
module. To prevent redundant codes, I returned the applogic module,
moved the getCategoriesRecords() method from the newincome module to the
applogic. I then added a parameter to the method which will be passed a
variable stating the type of records to fetch when invoked. Then, a
parameter of type income is supplied to the function call made by the
save button

newexpensegui module
In this module, I added an import statement for the Application Logic
module, invoked assigned the getCategoryRecords() to the combobox widget
and passed an argument of type expense to the method.

applogic module
As explained above, this module is to cater for codes that might be
redundant in other modules.

After all changes, the program executes without any errors