Method Implemenetation
This commit involves an implmenetation of a method to fetch records from
the database and display them in the treeview widget on the records gui.
So far, it executes without any error. Explained below is how it was
done and the observation I made:

In the recordsgui module, I added a method called displayRecords().
First, it has a variable assigned to the currently selected item of the
combobox widget to hold any selected statement. Using an if elif
statement, i assigned a value of Income or Expense to a variable, type,
depending on whether the user selected Income Categories or Expense
Categories. I tried only these two options for testing purposes. After
that, I invoked the getCategoryRecords() method from the applogic module
and passed the variable, type to it. Using a while loop with variable i
starting from zero and less than the length of the list from the result
returned by the getCategoryRecords() method, I invoked the set() method
of the variable assigned to the treeview widget and passed to its arguments
the values of i and variable holding the list from the
getCategoryRecords() method and incremented i by one each time round the
loop.

Observation
When the program runs, no error occurs and when either Income Categories
or Expense Categories is selected in the combobox widget, the list of
items of the selected type from the database get listed in the treeview
widget.

The only downside is that, when I select another option from the
combobox widget to display its list, it appends to the previous list in
the treeview widget. This behaviour, though it displays records without
error is not what I expected; I expected it to erase the previous lists
before displaying the next list and not append to it.