Module Improvement
In the immediate previous commit, I explained that I added a method that
when invoked by the click of a button displays a list of records of
category names in the treeview widget when an item is selected from the
combobox widget. Although the program executed without error, its
behaviour was unexpected. In this commit, I explain how I have fixed the
problem:

First, I created a new method, displayCategories() with an argument
variable, type. This variable will determine whether to display income
categories or expense categories. Then I cut and pasted the treeview
deifintion declared earlier in this method. This means that when this
method is invoked, it will define a new treeview widget with its column
and heading properties set. I then added a variable and assigned the
returned value from calling the getCategoryRecords() method from the
applogic module. I passed this method's argument, type to the
getCatrgoryRecords() method to ensure the correct records are fetched
from the database.

Finally, I added the while loop to display the list of records in the
treeview module. In the displayRecords() method created in the previous
commit, I deleted the while loop and rather called the
displayCategories() method and passed a variable holding the type of
categories to display.

However, to ensure that the gui display correctly, I defined a temporary
treeview widget which will be overriden when the displayRecords() method
is invoked by the click of the button widget.

Reason for this approach
Doing this solves the problem whereby different records get appended to
other records in the treeview. By this, only one type of record gets
displayed in the treeview widget.

Other minor tweaks were done for the program to execute without error.