Display Detail Records
In this commit, I explain how I got to display a detail list of records
of categories in the treeview widget

recordsgui module
In this module, I assigned the return value of get categories records
from the dbclass module to a variable. I then applied a for loop to the
records and added the record to the treeview variable.

When executed, the program runs without any error