Added New Methods
Added new methods to the dbclass and the recordsgui modules as explained
below:

dbclass module
In this module, I duplicated the _get_category_records() method and
renamed it to _get_transaction_records(). I then changed its second
parameter to _transaction_table which determines the database table to
fetch records from. In the select statement, I added the parameter to
it in a double plus (++) and single quotes to enable sql to read the
information.

recordsgui module
In this module, I duplicated the displayCategories() method and renamed
it to displayTransactions() with a parameter to specify the table name.
I then updated the displayRecords() to call the displayTransactions()
method depending on the item selected in the combobox widget.

The program was then executed and tested by selecting all options in turn
from the combobox widget. Each record gets displayed in the treeview
widget whenever the button clicked after each selection. One other to
add are scroll bars to the treeview to enable scrolling of records. This
will be added later after first phase of development.