Program Execution Test
I made some modifications and executed the program as explained below:

recordsgui module
In this module, I changed the if elif statements in the displayRecords()
method to match case statements.

mywallet.db file
All modules of the program were executed and data added which updated
the contents of the database file.

So far, all good.