New Method and Database Update
Added a new method and created a new database table as explained below:

database file
Created a new table, wallet_transactions to store records of all
transactions, be it of type income or expense.

dbclass module
In this module, I duplicated the _commitexpenses() method and renamed it
to _commitransactions(). I then updated the parameters,  variables and
the sql statement to connect to the newly created table in the database
as described above.

Purpose for this new additions
This new changes was done to remove redundant codes and tables; tables
that share similarities such as the income_records and expense_records
are put together into the transactions table. In the dbclass module, the
methods for committing income and expense transactions have been put
together into the commit transactions method.