Function Call Update
In the previous commit, I created a new table and a method to handle
collective transactions of both income and expense type.

In the newexpense module, I updated the function call for passing data
from the gui widgets to the database. I then executed the program to
test for its functionality.

The program executes without error.