Updated Database
Updated the database file as explained below:

mywallet.db file
Added a new column, transaction_type to the wallet_transactions table.
This column will store the type a particular transaction is associated
with, either incomes or expenses. The transaction_category column will
store data about the category of the transaction, such as Salary,
Clothing, Grocery etc.

Reason for this update
The transactions table storeis all data related to transactions. From
this table, data will be pulled to display different types of records,
including charts and graphs. The addition of the transaction_type to
flag whether a transaction is of type income or expense will make any
reporting type easy to handle.