Added Multiple Methods
Duplicated the get_number_of_records() method as explained below:

_get_number_of_records_of_types() method
Renamed the _get_number_of_records() to
_get_number_of_records_of_types() and changed its parameter to
_transaction_type. When invoked, this method computes for the number of
records of a specified type from the transactions table.

_get_sum_of_amounts_of_types() method
This method, when invoked will compute for the sum of amounts for all
records of the type specified by the parameter.

_get_number_of_records_of_categories() method
When invoked, this method computes for the number of records of a
specified category from the transactions table. The category is
specified by the passed parameter.

_get_sum_of_amounts_of_categories() method
This method computes for the sum of amounts for all records from the
transaction table for categories specified by the parameter when it is
invoked.